+++ b00t-cli/src/commands/cli_cmd.rs
@@ -1,0 +1,46 @@
+use anyhow::Result;
+use clap::Parser;
+
+#[derive(Parser)]
+pub enum CliCommands {
+    #[clap(
+        about = "Run a CLI script by name",
+        long_about = "Run a CLI script by name.\n\nExamples:\n  b00t-cli cli run setup-dev\n  b00t-cli cli run deploy"
+    )]
+    Run {
+        #[clap(help = "Script name to run")]
+        script_name: String,
+        #[clap(
+            help = "Arguments to pass to the script",
+            raw = true,
+            trailing_var_arg = true
+        )]
+        args: Vec<String>,
+    },
+}
+
+impl CliCommands {
+    pub fn execute(&self, _path: &str) -> Result<()> {
+        match self {
+            CliCommands::Run { .. } => {
+                println!("🚀 CLI run functionality coming soon...");
+                Ok(())
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_cli_commands_exist() {
+        let run_cmd = CliCommands::Run {
+            script_name: "test-script".to_string(),
+            args: vec![],
+        };
+        
+        assert!(run_cmd.execute("test").is_ok());
+    }
+}
