+++ b00t-cli/src/commands/init.rs
@@ -1,0 +1,33 @@
+use anyhow::Result;
+use clap::Parser;
+
+#[derive(Parser)]
+pub enum InitCommands {
+    #[clap(
+        about = "Initialize command aliases",
+        long_about = "Initialize command aliases for CLI tools.\n\nExamples:\n  b00t-cli init aliases"
+    )]
+    Aliases,
+}
+
+impl InitCommands {
+    pub fn execute(&self, _path: &str) -> Result<()> {
+        match self {
+            InitCommands::Aliases => {
+                println!("🔗 Init aliases functionality coming soon...");
+                Ok(())
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_init_commands_exist() {
+        let aliases_cmd = InitCommands::Aliases;
+        assert!(aliases_cmd.execute("test").is_ok());
+    }
+}
