+++ b00t-cli/src/commands/k8s.rs
@@ -1,0 +1,122 @@
+use anyhow::Result;
+use clap::Parser;
+
+#[derive(Parser)]
+pub enum K8sCommands {
+    #[clap(
+        about = "Deploy a pod from Dockerfile or docker-compose",
+        long_about = "Deploy a pod from Dockerfile or docker-compose.\n\nExamples:\n  b00t-cli k8s deploy --from-dockerfile ./Dockerfile --name web-server\n  b00t-cli k8s deploy --from-compose ./docker-compose.yaml\n  b00t-cli k8s deploy --image nginx:latest --name nginx-test"
+    )]
+    Deploy {
+        #[clap(long, help = "Deploy from Dockerfile", conflicts_with_all = &["from_compose", "image"])]
+        from_dockerfile: Option<String>,
+        #[clap(long, help = "Deploy from docker-compose.yaml", conflicts_with_all = &["from_dockerfile", "image"])]
+        from_compose: Option<String>,
+        #[clap(long, help = "Deploy from container image", conflicts_with_all = &["from_dockerfile", "from_compose"])]
+        image: Option<String>,
+        #[clap(long, help = "Pod name (required for dockerfile/image deployment)")]
+        name: Option<String>,
+        #[clap(long, help = "Namespace (default: default)")]
+        namespace: Option<String>,
+        #[clap(long, help = "Environment variables in KEY=VALUE format")]
+        env: Vec<String>,
+    },
+    #[clap(
+        about = "Deploy MCP server as Kubernetes pod",
+        long_about = "Deploy MCP server as Kubernetes pod.\n\nExamples:\n  b00t-cli k8s deploy-mcp --server filesystem\n  b00t-cli k8s deploy-mcp --server brave-search --namespace mcp-servers"
+    )]
+    DeployMcp {
+        #[clap(long, help = "MCP server name from b00t configuration")]
+        server: String,
+        #[clap(long, help = "Namespace (default: default)")]
+        namespace: Option<String>,
+        #[clap(long, help = "Override pod name")]
+        name: Option<String>,
+    },
+    #[clap(
+        about = "List running pods",
+        long_about = "List running pods managed by b00t.\n\nExamples:\n  b00t-cli k8s list\n  b00t-cli k8s list --namespace kube-system\n  b00t-cli k8s list --all"
+    )]
+    List {
+        #[clap(long, help = "Show pods in specific namespace")]
+        namespace: Option<String>,
+        #[clap(long, help = "Show all pods (not just b00t-managed)")]
+        all: bool,
+        #[clap(long, help = "Output in JSON format")]
+        json: bool,
+    },
+    #[clap(
+        about = "Show pod logs",
+        long_about = "Show logs for a specific pod.\n\nExamples:\n  b00t-cli k8s logs web-server\n  b00t-cli k8s logs --follow web-server\n  b00t-cli k8s logs --previous web-server"
+    )]
+    Logs {
+        #[clap(help = "Pod name")]
+        pod_name: String,
+        #[clap(long, help = "Namespace (default: default)")]
+        namespace: Option<String>,
+        #[clap(long, help = "Follow log output")]
+        follow: bool,
+        #[clap(long, help = "Show previous container logs")]
+        previous: bool,
+    },
+    #[clap(
+        about = "Delete resources",
+        long_about = "Delete Kubernetes resources.\n\nExamples:\n  b00t-cli k8s delete pod web-server\n  b00t-cli k8s delete --all pods\n  b00t-cli k8s delete service web-service"
+    )]
+    Delete {
+        #[clap(help = "Resource type (pod, service, deployment)")]
+        resource_type: String,
+        #[clap(help = "Resource name (or --all for all resources)")]
+        resource_name: Option<String>,
+        #[clap(long, help = "Delete all resources of the specified type")]
+        all: bool,
+        #[clap(long, help = "Namespace (default: default)")]
+        namespace: Option<String>,
+    },
+}
+
+impl K8sCommands {
+    pub fn execute(&self, _path: &str) -> Result<()> {
+        match self {
+            K8sCommands::Deploy { .. } => {
+                println!("🚀 K8s deploy functionality coming soon...");
+                Ok(())
+            }
+            K8sCommands::DeployMcp { .. } => {
+                println!("🚀 K8s deploy-mcp functionality coming soon...");
+                Ok(())
+            }
+            K8sCommands::List { .. } => {
+                println!("📋 K8s list functionality coming soon...");
+                Ok(())
+            }
+            K8sCommands::Logs { .. } => {
+                println!("📜 K8s logs functionality coming soon...");
+                Ok(())
+            }
+            K8sCommands::Delete { .. } => {
+                println!("🗑️ K8s delete functionality coming soon...");
+                Ok(())
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_k8s_commands_exist() {
+        let deploy_cmd = K8sCommands::Deploy {
+            from_dockerfile: Some("Dockerfile".to_string()),
+            from_compose: None,
+            image: None,
+            name: Some("test-pod".to_string()),
+            namespace: None,
+            env: vec![],
+        };
+        
+        assert!(deploy_cmd.execute("test").is_ok());
+    }
+}
