+++ b00t-cli/src/commands/mod.rs
@@ -1,0 +1,17 @@
+pub mod mcp;
+pub mod k8s;
+pub mod ai;
+pub mod app;
+pub mod cli_cmd;
+pub mod init;
+pub mod session;
+pub mod whatismy;
+
+pub use mcp::McpCommands;
+pub use k8s::K8sCommands;
+pub use ai::AiCommands;
+pub use app::AppCommands;
+pub use cli_cmd::CliCommands;
+pub use init::InitCommands;
+pub use session::SessionCommands;
+pub use whatismy::WhatismyCommands;
