+++ b00t-cli/src/commands/session.rs
@@ -1,0 +1,51 @@
+use anyhow::Result;
+use clap::Parser;
+
+#[derive(Parser)]
+pub enum SessionCommands {
+    #[clap(about = "Initialize new session")]
+    Init {
+        #[clap(long, help = "Budget limit in USD")]
+        budget: Option<f64>,
+        #[clap(long, help = "Session name")]
+        name: Option<String>,
+    },
+    #[clap(about = "Show current session status")]
+    Status,
+    #[clap(about = "End current session")]
+    End,
+}
+
+impl SessionCommands {
+    pub fn execute(&self, _path: &str) -> Result<()> {
+        match self {
+            SessionCommands::Init { .. } => {
+                println!("🆕 Session init functionality coming soon...");
+                Ok(())
+            }
+            SessionCommands::Status => {
+                println!("📊 Session status functionality coming soon...");
+                Ok(())
+            }
+            SessionCommands::End => {
+                println!("🔚 Session end functionality coming soon...");
+                Ok(())
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_session_commands_exist() {
+        let init_cmd = SessionCommands::Init {
+            budget: Some(10.0),
+            name: Some("test-session".to_string()),
+        };
+        
+        assert!(init_cmd.execute("test").is_ok());
+    }
+}
