+++ b00t-cli/src/commands/whatismy.rs
@@ -1,0 +1,57 @@
+use anyhow::Result;
+use clap::Parser;
+
+#[derive(Parser)]
+pub enum WhatismyCommands {
+    #[clap(about = "Detect current AI agent (claude, gemini, etc.)")]
+    Agent {
+        #[clap(long, help = "Ignore _B00T_Agent environment variable")]
+        no_env: bool,
+        #[clap(long, help = "Output in JSON format")]
+        json: bool,
+    },
+    #[clap(about = "Detect current session information")]
+    Session {
+        #[clap(long, help = "Output in JSON format")]
+        json: bool,
+    },
+    #[clap(about = "Detect current environment setup")]
+    Environment {
+        #[clap(long, help = "Output in JSON format")]
+        json: bool,
+    },
+}
+
+impl WhatismyCommands {
+    pub fn execute(&self, _path: &str) -> Result<()> {
+        match self {
+            WhatismyCommands::Agent { .. } => {
+                println!("🤖 Whatismy agent functionality coming soon...");
+                Ok(())
+            }
+            WhatismyCommands::Session { .. } => {
+                println!("📊 Whatismy session functionality coming soon...");
+                Ok(())
+            }
+            WhatismyCommands::Environment { .. } => {
+                println!("🌍 Whatismy environment functionality coming soon...");
+                Ok(())
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_whatismy_commands_exist() {
+        let agent_cmd = WhatismyCommands::Agent {
+            no_env: false,
+            json: false,
+        };
+        
+        assert!(agent_cmd.execute("test").is_ok());
+    }
+}
