+++ b00t-cli/src/main.rs
@@ -13,6 +13,7 @@ use b00t_cli::{
     normalize_mcp_json,
 };
 
+mod commands;
 mod datum_ai;
 mod datum_apt;
 mod datum_bash;
@@ -34,6 +35,11 @@ use datum_gemini::gemini_install_mcp;
 use datum_mcp::McpDatum;
 use datum_vscode::VscodeDatum;
 use traits::*;
+
+use crate::commands::{AiCommands, AppCommands, CliCommands, InitCommands, K8sCommands, McpCommands, SessionCommands, WhatismyCommands};
+
+// Re-export commonly used functions for datum modules
+pub use b00t_cli::{get_config, get_expanded_path, get_mcp_config, mcp_add_json, mcp_list};
 
 mod integration_tests;
 
@@ -114,430 +120,6 @@ enum Commands {
         #[clap(subcommand)]
         session_command: SessionCommands,
     },
-}
-
-#[derive(Parser)]
-enum McpCommands {
-    #[clap(
-        about = "Create MCP server configuration",
-        long_about = "Create MCP server configuration from JSON or command.\n\nJSON Examples:\n  b00t-cli mcp create '{\"name\":\"filesystem\",\"command\":\"npx\",\"args\":[\"-y\",\"@modelcontextprotocol/server-filesystem\"]}'\n  echo '{...}' | b00t-cli mcp create -\n\nCommand Examples:\n  b00t-cli mcp create brave-search -- npx -y @modelcontextprotocol/server-brave-search\n  b00t-cli mcp create filesystem --hint \"File system access\" -- npx -y @modelcontextprotocol/server-filesystem\n\nInstallation Examples:\n  b00t-cli mcp install brave-search claudecode\n  b00t-cli app vscode mcp install filesystem"
-    )]
-    Create {
-        #[clap(help = "MCP server name (for command mode) or JSON configuration (for JSON mode)")]
-        name_or_json: String,
-        #[clap(long, help = "Description/hint for the MCP server")]
-        hint: Option<String>,
-        #[clap(
-            long,
-            help = "Do What I Want - auto-cleanup and format JSON (default: enabled)"
-        )]
-        dwiw: bool,
-        #[clap(
-            long,
-            help = "Disable auto-cleanup and format JSON",
-            conflicts_with = "dwiw"
-        )]
-        no_dwiw: bool,
-        #[clap(
-            last = true,
-            help = "Command and arguments (after --) for command mode"
-        )]
-        command_args: Vec<String>,
-    },
-    #[clap(
-        about = "List available MCP server configurations",
-        long_about = "List available MCP server configurations.\n\nExamples:\n  b00t-cli mcp list\n  b00t-cli mcp list --json"
-    )]
-    List {
-        #[clap(long, help = "Output in JSON format")]
-        json: bool,
-    },
-    #[clap(
-        about = "Install MCP server to a target (claudecode, vscode, geminicli)",
-        long_about = "Install MCP server to a target application.\n\nExamples:\n  b00t-cli mcp install gh claudecode\n  b00t-cli mcp install filesystem geminicli --repo\n  b00t-cli app vscode mcp install filesystem"
-    )]
-    Install {
-        #[clap(help = "MCP server name")]
-        name: String,
-        #[clap(help = "Installation target: claudecode, vscode, geminicli")]
-        target: String,
-        #[clap(long, help = "Install to repository-specific location (for geminicli)")]
-        repo: bool,
-        #[clap(long, help = "Install to user-global location (for geminicli)")]
-        user: bool,
-    },
-    #[clap(
-        about = "Output MCP servers in various formats",
-        long_about = "Output MCP servers in various formats for configuration files.\n\nExamples:\n  b00t-cli mcp output filesystem,brave-search\n  b00t-cli mcp output --json filesystem\n  b00t-cli mcp output --mcpServers filesystem,brave-search"
-    )]
-    Output {
-        #[clap(long = "json", help = "Output raw JSON format without wrapper", action = clap::ArgAction::SetTrue)]
-        json: bool,
-        #[clap(long = "mcpServers", help = "Output in mcpServers format (default)", action = clap::ArgAction::SetTrue)]
-        mcp_servers: bool,
-        #[clap(help = "Comma-separated list of MCP server names to output")]
-        servers: String,
-    },
-}
-
-#[derive(Parser)]
-enum K8sCommands {
-    #[clap(
-        about = "Deploy a pod from Dockerfile or docker-compose",
-        long_about = "Deploy a pod from Dockerfile or docker-compose.\n\nExamples:\n  b00t-cli k8s deploy --from-dockerfile ./Dockerfile --name web-server\n  b00t-cli k8s deploy --from-compose ./docker-compose.yaml\n  b00t-cli k8s deploy --image nginx:latest --name nginx-test"
-    )]
-    Deploy {
-        #[clap(long, help = "Deploy from Dockerfile", conflicts_with_all = &["from_compose", "image"])]
-        from_dockerfile: Option<String>,
-        #[clap(long, help = "Deploy from docker-compose.yaml", conflicts_with_all = &["from_dockerfile", "image"])]
-        from_compose: Option<String>,
-        #[clap(long, help = "Deploy from container image", conflicts_with_all = &["from_dockerfile", "from_compose"])]
-        image: Option<String>,
-        #[clap(long, help = "Pod name (required for dockerfile/image deployment)")]
-        name: Option<String>,
-        #[clap(long, help = "Namespace (default: default)")]
-        namespace: Option<String>,
-        #[clap(long, help = "Environment variables in KEY=VALUE format")]
-        env: Vec<String>,
-    },
-    #[clap(
-        about = "Deploy MCP server as Kubernetes pod",
-        long_about = "Deploy MCP server as Kubernetes pod.\n\nExamples:\n  b00t-cli k8s deploy-mcp --server filesystem\n  b00t-cli k8s deploy-mcp --server brave-search --namespace mcp-servers"
-    )]
-    DeployMcp {
-        #[clap(long, help = "MCP server name from b00t configuration")]
-        server: String,
-        #[clap(long, help = "Namespace (default: default)")]
-        namespace: Option<String>,
-        #[clap(long, help = "Override pod name")]
-        name: Option<String>,
-    },
-    #[clap(
-        about = "List running pods",
-        long_about = "List running pods managed by b00t.\n\nExamples:\n  b00t-cli k8s list\n  b00t-cli k8s list --namespace kube-system\n  b00t-cli k8s list --all"
-    )]
-    List {
-        #[clap(long, help = "Show pods in specific namespace")]
-        namespace: Option<String>,
-        #[clap(long, help = "Show all pods (not just b00t-managed)")]
-        all: bool,
-        #[clap(long, help = "Output in JSON format")]
-        json: bool,
-    },
-    #[clap(
-        about = "Show pod logs",
-        long_about = "Show logs for a specific pod.\n\nExamples:\n  b00t-cli k8s logs web-server\n  b00t-cli k8s logs --follow web-server\n  b00t-cli k8s logs --previous web-server"
-    )]
-    Logs {
-        #[clap(help = "Pod name")]
-        pod_name: String,
-        #[clap(long, help = "Namespace (default: default)")]
-        namespace: Option<String>,
-        #[clap(short, long, help = "Follow log output")]
-        follow: bool,
-        #[clap(long, help = "Show previous container logs")]
-        previous: bool,
-        #[clap(long, help = "Number of lines to show")]
-        tail: Option<i64>,
-    },
-    #[clap(
-        about = "Delete a pod",
-        long_about = "Delete a pod.\n\nExamples:\n  b00t-cli k8s delete web-server\n  b00t-cli k8s delete --namespace test web-server"
-    )]
-    Delete {
-        #[clap(help = "Pod name")]
-        pod_name: String,
-        #[clap(long, help = "Namespace (default: default)")]
-        namespace: Option<String>,
-    },
-    #[clap(
-        about = "Restart an MCP server pod",
-        long_about = "Restart an MCP server pod.\n\nExamples:\n  b00t-cli k8s restart-mcp filesystem\n  b00t-cli k8s restart-mcp --namespace mcp-servers brave-search"
-    )]
-    RestartMcp {
-        #[clap(help = "MCP server name")]
-        server_name: String,
-        #[clap(long, help = "Namespace (default: default)")]
-        namespace: Option<String>,
-    },
-    #[clap(
-        about = "Show cluster status",
-        long_about = "Show Kubernetes cluster status.\n\nExamples:\n  b00t-cli k8s status\n  b00t-cli k8s status --detailed"
-    )]
-    Status {
-        #[clap(long, help = "Show detailed cluster information")]
-        detailed: bool,
-    },
-    #[clap(
-        about = "Setup minikube cluster",
-        long_about = "Setup and validate minikube cluster.\n\nExamples:\n  b00t-cli k8s setup\n  b00t-cli k8s setup --force"
-    )]
-    Setup {
-        #[clap(long, help = "Force cluster recreate if exists")]
-        force: bool,
-    },
-    #[clap(
-        about = "Teardown cluster resources",
-        long_about = "Teardown cluster or specific resources.\n\nExamples:\n  b00t-cli k8s teardown --all\n  b00t-cli k8s teardown --namespace test"
-    )]
-    Teardown {
-        #[clap(long, help = "Teardown entire cluster")]
-        all: bool,
-        #[clap(long, help = "Teardown specific namespace")]
-        namespace: Option<String>,
-    },
-}
-
-#[derive(Parser)]
-enum AiCommands {
-    #[clap(
-        about = "Add AI provider configuration from TOML file",
-        long_about = "Add AI provider configuration from TOML file.\n\nExamples:\n  b00t-cli ai add ./openai.ai.toml\n  b00t-cli ai add ~/.dotfiles/_b00t_/anthropic.ai.toml"
-    )]
-    Add {
-        #[clap(help = "Path to AI provider TOML file")]
-        file: String,
-    },
-    #[clap(
-        about = "List available AI provider configurations",
-        long_about = "List available AI provider configurations.\n\nExamples:\n  b00t-cli ai list\n  b00t-cli ai list --json"
-    )]
-    List {
-        #[clap(long, help = "Output in JSON format")]
-        json: bool,
-    },
-    #[clap(
-        about = "Output AI providers in various formats",
-        long_about = "Output AI providers in various formats.\n\nExamples:\n  b00t-cli ai output --kv openai,anthropic\n  b00t-cli ai output --b00t openai\n  b00t-cli ai output anthropic"
-    )]
-    Output {
-        #[clap(long = "b00t", help = "Output in b00t TOML format (default)", action = clap::ArgAction::SetTrue)]
-        b00t: bool,
-        #[clap(long = "kv", help = "Output environment variables in KEY=VALUE format", action = clap::ArgAction::SetTrue)]
-        kv: bool,
-        #[clap(help = "Comma-separated list of AI provider names to output")]
-        providers: String,
-    },
-}
-
-#[derive(Parser)]
-enum AppCommands {
-    #[clap(
-        about = "VSCode integration commands",
-        long_about = "VSCode integration commands.\n\nMCP Examples:\n  b00t-cli app vscode mcp install gh\n  b00t-cli mcp install gh vscode\n\nExtension Examples:\n  b00t-cli app vscode extension list\n  b00t-cli app vscode extension install rust-lang.rust-analyzer"
-    )]
-    Vscode {
-        #[clap(subcommand)]
-        vscode_command: AppVscodeCommands,
-    },
-    #[clap(
-        about = "Claude Code integration commands",
-        long_about = "Claude Code integration commands.\n\nExamples:\n  b00t-cli app claudecode mcp install gh\n  b00t-cli mcp install gh claudecode"
-    )]
-    Claudecode {
-        #[clap(subcommand)]
-        claudecode_command: AppClaudecodeCommands,
-    },
-    #[clap(
-        about = "Gemini CLI integration commands",
-        long_about = "Gemini CLI integration commands.\n\nExamples:\n  b00t-cli app geminicli mcp install gh --repo\n  b00t-cli mcp install gh geminicli --user"
-    )]
-    Geminicli {
-        #[clap(subcommand)]
-        geminicli_command: AppGeminicliCommands,
-    },
-}
-
-#[derive(Parser)]
-enum AppVscodeCommands {
-    #[clap(about = "MCP server management for VSCode")]
-    Mcp {
-        #[clap(subcommand)]
-        mcp_command: AppMcpCommands,
-    },
-    #[clap(about = "VS Code extension management")]
-    Extension {
-        #[clap(subcommand)]
-        extension_command: AppVscodeExtensionCommands,
-    },
-}
-
-#[derive(Parser)]
-enum AppClaudecodeCommands {
-    #[clap(about = "MCP server management for Claude Code")]
-    Mcp {
-        #[clap(subcommand)]
-        mcp_command: AppMcpCommands,
-    },
-}
-
-#[derive(Parser)]
-enum AppGeminicliCommands {
-    #[clap(about = "MCP server management for Gemini CLI")]
-    Mcp {
-        #[clap(subcommand)]
-        mcp_command: AppGeminicliMcpCommands,
-    },
-}
-
-#[derive(Parser)]
-enum AppMcpCommands {
-    #[clap(
-        about = "Install MCP server",
-        long_about = "Install MCP server to the target application.\n\nExamples:\n  b00t-cli app vscode mcp install gh\n  b00t-cli app claudecode mcp install filesystem"
-    )]
-    Install {
-        #[clap(help = "Name of the MCP server to install")]
-        name: String,
-    },
-}
-
-#[derive(Parser)]
-enum AppVscodeExtensionCommands {
-    #[clap(
-        about = "List installed VS Code extensions",
-        long_about = "List all installed VS Code extensions.\n\nExamples:\n  b00t-cli app vscode extension list\n  b00t-cli app vscode extension list --json"
-    )]
-    List {
-        #[clap(long, help = "Output in JSON format")]
-        json: bool,
-    },
-    #[clap(
-        about = "Install VS Code extension",
-        long_about = "Install VS Code extension by ID.\n\nExamples:\n  b00t-cli app vscode extension install ms-vscode.vscode-typescript-next\n  b00t-cli app vscode extension install rust-lang.rust-analyzer"
-    )]
-    Install {
-        #[clap(help = "Extension ID to install (e.g., ms-vscode.vscode-typescript-next)")]
-        extension_id: String,
-    },
-    #[clap(
-        about = "Uninstall VS Code extension",
-        long_about = "Uninstall VS Code extension by ID.\n\nExamples:\n  b00t-cli app vscode extension uninstall ms-vscode.vscode-typescript-next\n  b00t-cli app vscode extension uninstall rust-lang.rust-analyzer"
-    )]
-    Uninstall {
-        #[clap(help = "Extension ID to uninstall")]
-        extension_id: String,
-    },
-}
-
-#[derive(Parser)]
-enum AppGeminicliMcpCommands {
-    #[clap(
-        about = "Install MCP server to Gemini CLI",
-        long_about = "Install MCP server to Gemini CLI extension.\n\nExamples:\n  b00t-cli app geminicli mcp install gh --repo\n  b00t-cli app geminicli mcp install filesystem --user"
-    )]
-    Install {
-        #[clap(help = "Name of the MCP server to install")]
-        name: String,
-        #[clap(
-            long,
-            help = "Install to repository-specific extension (default if in git repo)"
-        )]
-        repo: bool,
-        #[clap(long, help = "Install to user-global extension")]
-        user: bool,
-    },
-}
-
-#[derive(Parser)]
-enum CliCommands {
-    #[clap(
-        about = "Run a CLI script by name",
-        long_about = "Run a CLI script by name.\n\nExamples:\n  b00t-cli cli run setup-dev\n  b00t-cli cli run deploy"
-    )]
-    Run {
-        #[clap(help = "Name of the CLI script to run")]
-        name: String,
-    },
-    #[clap(
-        about = "Detect the version of a CLI command",
-        long_about = "Detect the version of a CLI command.\n\nExamples:\n  b00t-cli cli detect node\n  b00t-cli cli detect git"
-    )]
-    Detect {
-        #[clap(help = "The command to detect")]
-        command: String,
-    },
-    #[clap(
-        about = "Show the desired version of a CLI command",
-        long_about = "Show the desired version of a CLI command.\n\nExamples:\n  b00t-cli cli desires node\n  b00t-cli cli desires docker"
-    )]
-    Desires {
-        #[clap(help = "The command to check")]
-        command: String,
-    },
-    #[clap(
-        about = "Install a CLI command",
-        long_about = "Install a CLI command using its configuration.\n\nExamples:\n  b00t-cli cli install node\n  b00t-cli cli install docker"
-    )]
-    Install {
-        #[clap(help = "The command to install")]
-        command: String,
-    },
-    #[clap(
-        about = "Update a CLI command",
-        long_about = "Update a CLI command using its configuration.\n\nExamples:\n  b00t-cli cli update node\n  b00t-cli cli update docker"
-    )]
-    Update {
-        #[clap(help = "The command to update")]
-        command: String,
-    },
-    #[clap(
-        about = "Check installed vs desired versions for CLI command",
-        long_about = "Check installed vs desired versions for CLI command.\n\nExamples:\n  b00t-cli cli check node\n  b00t-cli cli check docker\n\nOutput emojis:\n  🥾👍🏻 = versions match\n  🥾🐣 = installed version newer than desired\n  🥾😭 = installed version older than desired\n  🥾😱 = command missing"
-    )]
-    Check {
-        #[clap(help = "The command to check")]
-        command: String,
-    },
-    #[clap(
-        about = "Update all CLI commands",
-        long_about = "Update all CLI commands that have outdated versions.\n\nExamples:\n  b00t-cli cli up"
-    )]
-    Up,
-}
-
-#[derive(Parser)]
-enum InitCommands {
-    #[clap(
-        about = "Initialize command aliases",
-        long_about = "Initialize command aliases for CLI tools.\n\nExamples:\n  b00t-cli init aliases"
-    )]
-    Aliases,
-}
-
-#[derive(Parser)]
-enum WhatismyCommands {
-    #[clap(about = "Detect current AI agent (claude, gemini, etc.)")]
-    Agent {
-        #[clap(long, help = "Ignore _B00T_Agent environment variable")]
-        no_env: bool,
-    },
-}
-
-#[derive(Parser)]
-enum SessionCommands {
-    #[clap(about = "Initialize new session")]
-    Init {
-        #[clap(long, help = "Budget limit in USD")]
-        budget: Option<f64>,
-        #[clap(long, help = "Time limit in minutes")]
-        time_limit: Option<u32>,
-        #[clap(long, help = "Agent name override")]
-        agent: Option<String>,
-    },
-    #[clap(about = "Show current session status")]
-    Status,
-    #[clap(about = "Update session metrics")]
-    Update {
-        #[clap(long, help = "Estimated cost to add")]
-        cost: Option<f64>,
-        #[clap(long, help = "Hint message to add")]
-        hint: Option<String>,
-    },
-    #[clap(about = "End current session")]
-    End,
-    #[clap(about = "Get compact status for prompt")]
-    Prompt,
 }
 
 // Using unified config from lib.rs
@@ -1413,1857 +995,75 @@ fn main() {
     }
 
     match &cli.command {
-        Some(Commands::Mcp { mcp_command }) => match mcp_command {
-            McpCommands::Create {
-                name_or_json,
-                hint,
-                dwiw: _,
-                no_dwiw,
-                command_args,
-            } => {
-                if !command_args.is_empty() {
-                    // Command mode: b00t-cli mcp create name -- command args...
-                    if let Err(e) =
-                        mcp_add_command(name_or_json, hint.as_deref(), command_args, &cli.path)
-                    {
-                        eprintln!("Error adding MCP server: {}", e);
-                        std::process::exit(1);
-                    }
-                } else {
-                    // JSON mode: b00t-cli mcp create '{"name":"..."}'
-                    let use_dwiw = if *no_dwiw { false } else { true };
-                    if let Err(e) = mcp_add_json(name_or_json, use_dwiw, &cli.path) {
-                        eprintln!("Error adding MCP server: {}", e);
-                        std::process::exit(1);
-                    }
-                }
+        Some(Commands::Mcp { mcp_command }) => {
+            if let Err(e) = mcp_command.execute(&cli.path) {
+                eprintln!("Error: {}", e);
+                std::process::exit(1);
             }
-            McpCommands::List { json } => {
-                if let Err(e) = mcp_list(&cli.path, *json) {
-                    eprintln!("Error listing MCP servers: {}", e);
-                    std::process::exit(1);
-                }
+        }
+        Some(Commands::Ai { ai_command }) => {
+            if let Err(e) = ai_command.execute(&cli.path) {
+                eprintln!("Error: {}", e);
+                std::process::exit(1);
             }
-            McpCommands::Install {
-                name,
-                target,
-                repo,
-                user,
-            } => {
-                let result = match target.as_str() {
-                    "claudecode" => claude_code_install_mcp(name, &cli.path),
-                    "vscode" => vscode_install_mcp(name, &cli.path),
-                    "geminicli" => {
-                        // Determine installation location: default to repo if in git repo, otherwise user
-                        let use_repo = if *repo && *user {
-                            eprintln!("Error: Cannot specify both --repo and --user flags");
-                            std::process::exit(1);
-                        } else if *repo {
-                            true
-                        } else if *user {
-                            false
-                        } else {
-                            // Default behavior: repo if in git repo, otherwise user
-                            crate::utils::is_git_repo()
-                        };
-                        gemini_install_mcp(name, &cli.path, use_repo)
-                    }
-                    _ => {
-                        eprintln!(
-                            "Error: Invalid target '{}'. Valid targets are: claudecode, vscode, geminicli",
-                            target
-                        );
-                        std::process::exit(1);
-                    }
-                };
-                if let Err(e) = result {
-                    eprintln!("Error installing MCP server to {}: {}", target, e);
-                    std::process::exit(1);
-                }
+        }
+        Some(Commands::App { app_command }) => {
+            if let Err(e) = app_command.execute(&cli.path) {
+                eprintln!("Error: {}", e);
+                std::process::exit(1);
             }
-            McpCommands::Output {
-                json,
-                mcp_servers: _,
-                servers,
-            } => {
-                // Default to mcpServers format unless --json is specified
-                let use_wrapper = if *json {
-                    false
-                } else {
-                    true // Default behavior or explicit --mcpServers
-                };
-
-                if let Err(e) = mcp_output(&cli.path, use_wrapper, servers) {
-                    eprintln!("Error outputting MCP servers: {}", e);
-                    std::process::exit(1);
-                }
+        }
+        Some(Commands::Cli { cli_command }) => {
+            if let Err(e) = cli_command.execute(&cli.path) {
+                eprintln!("Error: {}", e);
+                std::process::exit(1);
             }
-        },
-        Some(Commands::Ai { ai_command }) => match ai_command {
-            AiCommands::Add { file } => {
-                if let Err(e) = ai_add(file, &cli.path) {
-                    eprintln!("Error adding AI provider: {}", e);
-                    std::process::exit(1);
-                }
+        }
+        Some(Commands::Init { init_command }) => {
+            if let Err(e) = init_command.execute(&cli.path) {
+                eprintln!("Error: {}", e);
+                std::process::exit(1);
             }
-            AiCommands::List { json } => {
-                if let Err(e) = ai_list(&cli.path, *json) {
-                    eprintln!("Error listing AI providers: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            AiCommands::Output {
-                b00t: _,
-                kv,
-                providers,
-            } => {
-                let format = if *kv { "kv" } else { "b00t" };
-                if let Err(e) = ai_output(&cli.path, format, providers) {
-                    eprintln!("Error outputting AI providers: {}", e);
-                    std::process::exit(1);
-                }
-            }
-        },
-        Some(Commands::App { app_command }) => match app_command {
-            AppCommands::Vscode { vscode_command } => match vscode_command {
-                AppVscodeCommands::Mcp { mcp_command } => match mcp_command {
-                    AppMcpCommands::Install { name } => {
-                        if let Err(e) = vscode_install_mcp(name, &cli.path) {
-                            eprintln!("Error installing MCP server to VSCode: {}", e);
-                            std::process::exit(1);
-                        }
-                    }
-                },
-                AppVscodeCommands::Extension { extension_command } => match extension_command {
-                    AppVscodeExtensionCommands::List { json } => {
-                        if let Err(e) = vscode_extension_list(*json) {
-                            eprintln!("Error listing VS Code extensions: {}", e);
-                            std::process::exit(1);
-                        }
-                    }
-                    AppVscodeExtensionCommands::Install { extension_id } => {
-                        if let Err(e) = vscode_extension_install(extension_id) {
-                            eprintln!("Error installing VS Code extension: {}", e);
-                            std::process::exit(1);
-                        }
-                    }
-                    AppVscodeExtensionCommands::Uninstall { extension_id } => {
-                        if let Err(e) = vscode_extension_uninstall(extension_id) {
-                            eprintln!("Error uninstalling VS Code extension: {}", e);
-                            std::process::exit(1);
-                        }
-                    }
-                },
-            },
-            AppCommands::Claudecode { claudecode_command } => match claudecode_command {
-                AppClaudecodeCommands::Mcp { mcp_command } => match mcp_command {
-                    AppMcpCommands::Install { name } => {
-                        if let Err(e) = claude_code_install_mcp(name, &cli.path) {
-                            eprintln!("Error installing MCP server to Claude Code: {}", e);
-                            std::process::exit(1);
-                        }
-                    }
-                },
-            },
-            AppCommands::Geminicli { geminicli_command } => match geminicli_command {
-                AppGeminicliCommands::Mcp { mcp_command } => match mcp_command {
-                    AppGeminicliMcpCommands::Install { name, repo, user } => {
-                        // Determine installation location: default to repo if in git repo, otherwise user
-                        let use_repo = if *repo && *user {
-                            eprintln!("Error: Cannot specify both --repo and --user flags");
-                            std::process::exit(1);
-                        } else if *repo {
-                            true
-                        } else if *user {
-                            false
-                        } else {
-                            // Default behavior: repo if in git repo, otherwise user
-                            crate::utils::is_git_repo()
-                        };
-
-                        if let Err(e) = gemini_install_mcp(name, &cli.path, use_repo) {
-                            eprintln!("Error installing MCP server to Gemini CLI: {}", e);
-                            std::process::exit(1);
-                        }
-                    }
-                },
-            },
-        },
-        Some(Commands::Cli { cli_command }) => match cli_command {
-            CliCommands::Run { name } => {
-                if let Err(e) = cli_run(name, &cli.path) {
-                    eprintln!("Error running CLI script: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            CliCommands::Detect { command } => cli_detect(command, &cli.path),
-            CliCommands::Desires { command } => cli_desires(command, &cli.path),
-            CliCommands::Install { command } => cli_install(command, &cli.path),
-            CliCommands::Update { command } => cli_update(command, &cli.path),
-            CliCommands::Check { command } => cli_check(command, &cli.path),
-            CliCommands::Up => cli_up(&cli.path),
-        },
-        Some(Commands::Init { init_command }) => match init_command {
-            InitCommands::Aliases => {
-                println!("Aliases initialization not yet implemented.");
-                println!(
-                    "This will scan for CLI tools with 'aliases' field and create ~/.local/bin scripts."
-                );
-            }
-        },
+        }
         Some(Commands::Whoami) => {
             if let Err(e) = whoami(&cli.path) {
-                eprintln!("Error displaying agent identity: {}", e);
+                eprintln!("Error: {}", e);
                 std::process::exit(1);
             }
         }
-        Some(Commands::Checkpoint {
-            message,
-            skip_tests,
-        }) => {
+        Some(Commands::Checkpoint { message, skip_tests }) => {
             if let Err(e) = checkpoint(message.as_deref(), *skip_tests) {
-                eprintln!("Error creating checkpoint: {}", e);
+                eprintln!("Error: {}", e);
                 std::process::exit(1);
             }
         }
-        Some(Commands::Whatismy { whatismy_command }) => match whatismy_command {
-            WhatismyCommands::Agent { no_env } => {
-                let agent = detect_agent(*no_env);
-                println!("{}", agent);
+        Some(Commands::Whatismy { whatismy_command }) => {
+            if let Err(e) = whatismy_command.execute(&cli.path) {
+                eprintln!("Error: {}", e);
+                std::process::exit(1);
             }
-        },
-        Some(Commands::Status {
-            filter,
-            installed,
-            available,
-        }) => {
+        }
+        Some(Commands::Status { filter, installed, available }) => {
             if let Err(e) = show_status(&cli.path, filter.as_deref(), *installed, *available) {
-                eprintln!("Error displaying status: {}", e);
+                eprintln!("Error: {}", e);
                 std::process::exit(1);
             }
         }
-        Some(Commands::K8s { k8s_command }) => match k8s_command {
-            K8sCommands::Deploy {
-                from_dockerfile,
-                from_compose,
-                image,
-                name,
-                namespace,
-                env,
-            } => {
-                if let Err(e) = k8s_deploy(
-                    from_dockerfile.as_deref(),
-                    from_compose.as_deref(),
-                    image.as_deref(),
-                    name.as_deref(),
-                    namespace.as_deref(),
-                    env,
-                    &cli.path,
-                ) {
-                    eprintln!("Error deploying to k8s: {}", e);
-                    std::process::exit(1);
-                }
+        Some(Commands::K8s { k8s_command }) => {
+            if let Err(e) = k8s_command.execute(&cli.path) {
+                eprintln!("Error: {}", e);
+                std::process::exit(1);
             }
-            K8sCommands::DeployMcp {
-                server,
-                namespace,
-                name,
-            } => {
-                if let Err(e) =
-                    k8s_deploy_mcp(server, namespace.as_deref(), name.as_deref(), &cli.path)
-                {
-                    eprintln!("Error deploying MCP server to k8s: {}", e);
-                    std::process::exit(1);
-                }
+        }
+        Some(Commands::Session { session_command }) => {
+            if let Err(e) = session_command.execute(&cli.path) {
+                eprintln!("Error: {}", e);
+                std::process::exit(1);
             }
-            K8sCommands::List {
-                namespace,
-                all,
-                json,
-            } => {
-                if let Err(e) = k8s_list(namespace.as_deref(), *all, *json, &cli.path) {
-                    eprintln!("Error listing k8s pods: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            K8sCommands::Logs {
-                pod_name,
-                namespace,
-                follow,
-                previous,
-                tail,
-            } => {
-                if let Err(e) = k8s_logs(
-                    pod_name,
-                    namespace.as_deref(),
-                    *follow,
-                    *previous,
-                    *tail,
-                    &cli.path,
-                ) {
-                    eprintln!("Error getting k8s logs: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            K8sCommands::Delete {
-                pod_name,
-                namespace,
-            } => {
-                if let Err(e) = k8s_delete(pod_name, namespace.as_deref(), &cli.path) {
-                    eprintln!("Error deleting k8s pod: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            K8sCommands::RestartMcp {
-                server_name,
-                namespace,
-            } => {
-                if let Err(e) = k8s_restart_mcp(server_name, namespace.as_deref(), &cli.path) {
-                    eprintln!("Error restarting MCP server: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            K8sCommands::Status { detailed } => {
-                if let Err(e) = k8s_status(*detailed, &cli.path) {
-                    eprintln!("Error getting k8s status: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            K8sCommands::Setup { force } => {
-                if let Err(e) = k8s_setup(*force, &cli.path) {
-                    eprintln!("Error setting up k8s cluster: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            K8sCommands::Teardown { all, namespace } => {
-                if let Err(e) = k8s_teardown(*all, namespace.as_deref(), &cli.path) {
-                    eprintln!("Error tearing down k8s resources: {}", e);
-                    std::process::exit(1);
-                }
-            }
-        },
-        Some(Commands::Session { session_command }) => match session_command {
-            SessionCommands::Init {
-                budget,
-                time_limit,
-                agent,
-            } => {
-                if let Err(e) = handle_session_init(budget, time_limit, agent.as_deref()) {
-                    eprintln!("Error initializing session: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            SessionCommands::Status => {
-                if let Err(e) = handle_session_status() {
-                    eprintln!("Error getting session status: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            SessionCommands::Update { cost, hint } => {
-                if let Err(e) = handle_session_update(cost, hint.as_deref()) {
-                    eprintln!("Error updating session: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            SessionCommands::End => {
-                if let Err(e) = handle_session_end() {
-                    eprintln!("Error ending session: {}", e);
-                    std::process::exit(1);
-                }
-            }
-            SessionCommands::Prompt => {
-                if let Err(e) = handle_session_prompt() {
-                    eprintln!("Error getting session prompt: {}", e);
-                    std::process::exit(1);
-                }
-            }
-        },
+        }
         None => {
             eprintln!("No command provided. Use --help for usage information.");
             std::process::exit(1);
         }
     }
-}
-
-#[allow(dead_code)]
-fn get_config(command: &str, path: &str) -> Result<(Config, String), Box<dyn std::error::Error>> {
-    // Try different file extensions in order of preference
-    let extensions = [
-        ".cli.toml",
-        ".mcp.toml",
-        ".vscode.toml",
-        ".docker.toml",
-        ".apt.toml",
-        ".nix.toml",
-        ".bash.toml",
-        ".toml",
-    ];
-
-    let mut path_buf = PathBuf::new();
-    path_buf.push(shellexpand::tilde(path).to_string());
-
-    for ext in &extensions {
-        let filename = format!("{}{}", command, ext);
-        path_buf.set_file_name(&filename);
-        if path_buf.exists() {
-            let content = fs::read_to_string(&path_buf)?;
-            let config: Config = toml::from_str(&content)?;
-            return Ok((config, filename));
-        }
-    }
-
-    eprintln!("{} UNDEFINED", command);
-    std::process::exit(100);
-}
-
-fn get_cli_unified_config(
-    command: &str,
-    path: &str,
-) -> Result<(Config, String), Box<dyn std::error::Error>> {
-    // Only look for CLI-specific files
-    let extensions = [".cli.toml"];
-
-    let base_path = PathBuf::from(shellexpand::tilde(path).to_string());
-
-    for ext in &extensions {
-        let filename = format!("{}{}", command, ext);
-        let full_path = base_path.join(&filename);
-        if full_path.exists() {
-            let content = fs::read_to_string(&full_path)?;
-            let config: Config = toml::from_str(&content)?;
-            return Ok((config, filename));
-        }
-    }
-
-    eprintln!("CLI package {} UNDEFINED", command);
-    std::process::exit(100);
-}
-
-#[allow(dead_code)]
-fn detect(command: &str, path: &str) {
-    if let Some(version) = get_installed_version(command, path) {
-        println!("{}", version);
-    } else {
-        eprintln!("🥾😱 {} MISSING", command);
-        std::process::exit(2);
-    }
-}
-
-#[allow(dead_code)]
-fn desires(command: &str, path: &str) {
-    match get_config(command, path) {
-        Ok((config, _)) => {
-            println!("{}", config.b00t.desires.as_deref().unwrap_or("N/A"));
-        }
-        Err(e) => eprintln!("Error reading config for {}: {}", command, e),
-    }
-}
-
-#[allow(dead_code)]
-fn install(command: &str, path: &str) {
-    match get_config(command, path) {
-        Ok((config, filename)) => match config.b00t.get_datum_type(Some(&filename)) {
-            DatumType::Unknown => {
-                if let Some(install_cmd) = &config.b00t.install {
-                    println!("Installing {}...", command);
-                    let result = cmd!("bash", "-c", install_cmd).run();
-                    match result {
-                        Ok(_) => println!("Installation complete."),
-                        Err(e) => eprintln!("Installation failed: {}", e),
-                    }
-                } else {
-                    eprintln!("No install command defined for {}", command);
-                }
-            }
-            DatumType::Mcp => {
-                eprintln!(
-                    "Use 'b00t-cli vscode install mcp {}' or 'b00t-cli claude-code install mcp {}' instead",
-                    command, command
-                );
-            }
-            _ => {
-                eprintln!(
-                    "Installation not yet supported for {:?} packages",
-                    config.b00t.get_datum_type(Some(&filename))
-                );
-            }
-        },
-        Err(e) => eprintln!("Error reading config for {}: {}", command, e),
-    }
-}
-
-#[allow(dead_code)]
-fn update(command: &str, path: &str) {
-    match get_config(command, path) {
-        Ok((config, filename)) => match config.b00t.get_datum_type(Some(&filename)) {
-            DatumType::Unknown => {
-                if let Some(update_cmd) =
-                    config.b00t.update.as_ref().or(config.b00t.install.as_ref())
-                {
-                    println!("Updating {}...", command);
-                    let result = cmd!("bash", "-c", update_cmd).run();
-                    match result {
-                        Ok(_) => println!("Update complete."),
-                        Err(e) => eprintln!("Update failed: {}", e),
-                    }
-                } else {
-                    eprintln!("No update or install command defined for {}", command);
-                }
-            }
-            _ => {
-                eprintln!(
-                    "Update not yet supported for {:?} packages",
-                    config.b00t.get_datum_type(Some(&filename))
-                );
-            }
-        },
-        Err(e) => eprintln!("Error reading config for {}: {}", command, e),
-    }
-}
-
-#[allow(dead_code)]
-fn get_installed_version(command: &str, path: &str) -> Option<String> {
-    if let Ok((config, _)) = get_config(command, path) {
-        if let Some(version_cmd) = &config.b00t.version {
-            if let Ok(output) = cmd!("bash", "-c", version_cmd).read() {
-                let re = Regex::new(
-                    config
-                        .b00t
-                        .version_regex
-                        .as_deref()
-                        .unwrap_or("\\d+\\.\\d+\\.\\d+"),
-                )
-                .unwrap();
-                if let Some(caps) = re.captures(&output) {
-                    return Some(caps[0].to_string());
-                }
-            }
-        }
-    }
-    None
-}
-
-#[allow(dead_code)]
-fn check(command: &str, path: &str) {
-    let desired_version_str = match get_config(command, path) {
-        Ok((config, _)) => config.b00t.desires.clone().unwrap_or_else(|| {
-            eprintln!("No desired version specified for {}", command);
-            std::process::exit(1);
-        }),
-        Err(e) => {
-            eprintln!("Error reading config for {}: {}", command, e);
-            std::process::exit(1);
-        }
-    };
-
-    let desired_version = Version::parse(&desired_version_str).unwrap();
-
-    if let Some(installed_version_str) = get_installed_version(command, path) {
-        let installed_version = Version::parse(&installed_version_str).unwrap();
-
-        if installed_version == desired_version {
-            println!("🥾👍🏻 {}", command);
-            std::process::exit(0);
-        } else if installed_version > desired_version {
-            println!("🥾🐣 {} {}", command, installed_version);
-            std::process::exit(0);
-        } else {
-            println!(
-                "🥾😭 {} IS {} WANTS {}",
-                command, installed_version, desired_version
-            );
-            std::process::exit(1);
-        }
-    } else {
-        println!("🥾😱 {} MISSING", command);
-        std::process::exit(2);
-    }
-}
-
-#[allow(dead_code)]
-fn up(path: &str) {
-    let expanded_path = shellexpand::tilde(path).to_string();
-    let entries = match fs::read_dir(&expanded_path) {
-        Ok(entries) => entries,
-        Err(e) => {
-            eprintln!("Error reading directory {}: {}", &expanded_path, e);
-            return;
-        }
-    };
-
-    for entry in entries {
-        if let Ok(entry) = entry {
-            let entry_path = entry.path();
-            if entry_path.extension().and_then(|s| s.to_str()) == Some("toml") {
-                if let Some(command) = entry_path.file_stem().and_then(|s| s.to_str()) {
-                    let desired_version_str = match get_config(command, path) {
-                        Ok((config, _)) => match config.b00t.desires.clone() {
-                            Some(version) => version,
-                            None => continue,
-                        },
-                        Err(_) => continue,
-                    };
-
-                    let desired_version = Version::parse(&desired_version_str).unwrap();
-
-                    if let Some(installed_version_str) = get_installed_version(command, path) {
-                        let installed_version = Version::parse(&installed_version_str).unwrap();
-
-                        if installed_version < desired_version {
-                            println!("Updating {}...", command);
-                            update(command, path);
-                        }
-                    } else {
-                        println!("Installing {}...", command);
-                        install(command, path);
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn create_mcp_toml_config(datum: &BootDatum, path: &str) -> Result<()> {
-    create_unified_toml_config(datum, path)
-}
-
-fn is_whitelisted_package_manager(command: &str) -> bool {
-    matches!(
-        command,
-        "bash" | "just" | "npx" | "uvx" | "pnpm" | "bunx" | "docker"
-    )
-}
-
-fn mcp_add_command(
-    name: &str,
-    hint: Option<&str>,
-    command_args: &[String],
-    path: &str,
-) -> Result<()> {
-    if command_args.is_empty() {
-        anyhow::bail!("No command provided. Use -- followed by the command and arguments.");
-    }
-
-    let command = &command_args[0];
-    if !is_whitelisted_package_manager(command) {
-        anyhow::bail!(
-            "Package manager '{}' is not whitelisted. Allowed: just, npx, uvx, pnpm, bunx, docker",
-            command
-        );
-    }
-
-    // Handle special cases for package managers
-    let (actual_command, args) = match command.as_str() {
-        "pnpm" => {
-            // For pnpm, we expect "pnpm dlx ..." so we need to validate "dlx" is present
-            if command_args.len() < 2 || command_args[1] != "dlx" {
-                anyhow::bail!("pnpm must be used with 'dlx' subcommand: pnpm dlx <package>");
-            }
-            ("pnpm".to_string(), command_args[1..].to_vec())
-        }
-        _ => (command.clone(), command_args[1..].to_vec()),
-    };
-
-    let datum = BootDatum {
-        name: name.to_string(),
-        datum_type: Some(DatumType::Mcp),
-        desires: None,
-        hint: hint.unwrap_or("MCP server").to_string(),
-        install: None,
-        update: None,
-        version: None,
-        version_regex: None,
-        command: Some(actual_command),
-        args: Some(args),
-        vsix_id: None,
-        script: None,
-        image: None,
-        docker_args: None,
-        oci_uri: None,
-        resource_path: None,
-        chart_path: None,
-        namespace: None,
-        values_file: None,
-        keywords: None,
-        package_name: None,
-        env: None,
-        require: None,
-        aliases: None,
-    };
-
-    create_mcp_toml_config(&datum, path)?;
-
-    println!("MCP server '{}' configuration saved.", datum.name);
-    println!(
-        "Command: {} {}",
-        datum.command.as_ref().unwrap(),
-        datum.args.as_ref().unwrap().join(" ")
-    );
-    println!(
-        "To install to VSCode: b00t-cli vscode install mcp {}",
-        datum.name
-    );
-    println!(
-        "To install to Claude Code: b00t-cli claude-code install mcp {}",
-        datum.name
-    );
-
-    Ok(())
-}
-
-fn mcp_add_json(json: &str, dwiw: bool, path: &str) -> Result<()> {
-    let json_content = if json == "-" {
-        let mut buffer = String::new();
-        match io::stdin().read_to_string(&mut buffer) {
-            Ok(_) => {
-                let trimmed = buffer.trim();
-                if trimmed.is_empty() {
-                    anyhow::bail!(
-                        "No input provided. Pipe JSON content or press Ctrl+D after pasting."
-                    );
-                }
-                trimmed.to_string()
-            }
-            Err(e) => {
-                anyhow::bail!(
-                    "Failed to read from stdin: {}. Pipe JSON content or use Ctrl+D after input.",
-                    e
-                );
-            }
-        }
-    } else {
-        json.trim().to_string()
-    };
-
-    let datum = normalize_mcp_json(&json_content, dwiw)?;
-
-    create_mcp_toml_config(&datum, path)?;
-
-    println!("MCP server '{}' configuration saved.", datum.name);
-    println!(
-        "To install to VSCode: b00t-cli vscode install mcp {}",
-        datum.name
-    );
-
-    Ok(())
-}
-
-fn get_expanded_path(path: &str) -> Result<PathBuf> {
-    Ok(PathBuf::from(shellexpand::tilde(path).to_string()))
-}
-
-fn get_mcp_toml_files(path: &str) -> Result<Vec<String>> {
-    let expanded_path = get_expanded_path(path)?;
-    let entries = fs::read_dir(&expanded_path)
-        .with_context(|| format!("Error reading directory {}", expanded_path.display()))?;
-
-    let mut mcp_files = Vec::new();
-    for entry in entries {
-        if let Ok(entry) = entry {
-            let entry_path = entry.path();
-            if let Some(file_name) = entry_path.file_name().and_then(|s| s.to_str()) {
-                if file_name.ends_with(".mcp.toml") {
-                    if let Some(server_name) = file_name.strip_suffix(".mcp.toml") {
-                        mcp_files.push(server_name.to_string());
-                    }
-                }
-            }
-        }
-    }
-    Ok(mcp_files)
-}
-
-fn mcp_list(path: &str, json_output: bool) -> Result<()> {
-    let mcp_files = get_mcp_toml_files(path)?;
-    let mut mcp_items = Vec::new();
-
-    for server_name in mcp_files {
-        match get_mcp_config(&server_name, path) {
-            Ok(datum) => {
-                mcp_items.push(McpListItem {
-                    name: server_name,
-                    command: datum.command.clone(),
-                    args: datum.args.clone(),
-                    hint: Some(datum.hint.clone()),
-                    error: None,
-                });
-            }
-            Err(e) => {
-                mcp_items.push(McpListItem {
-                    name: server_name,
-                    command: None,
-                    args: None,
-                    hint: None,
-                    error: Some(e.to_string()),
-                });
-            }
-        }
-    }
-
-    if json_output {
-        let expanded_path = get_expanded_path(path)?;
-        let output = McpListOutput {
-            servers: mcp_items,
-            path: expanded_path.display().to_string(),
-        };
-        let json_str = serde_json::to_string_pretty(&output)
-            .context("Failed to serialize MCP list to JSON")?;
-        println!("{}", json_str);
-    } else {
-        let expanded_path = get_expanded_path(path)?;
-        if mcp_items.is_empty() {
-            println!(
-                "No MCP server configurations found in {}",
-                expanded_path.display()
-            );
-            println!("Use 'b00t-cli mcp add <json>' to add MCP server configurations.");
-        } else {
-            println!("Available MCP servers in {}:", expanded_path.display());
-            println!();
-            for item in mcp_items {
-                match (&item.command, &item.args) {
-                    (Some(command), Some(args)) => {
-                        println!("📋 {} ({})", item.name, command);
-                        if !args.is_empty() {
-                            println!("   args: {}", args.join(" "));
-                        }
-                    }
-                    _ => {
-                        println!("❌ {} (error reading config)", item.name);
-                    }
-                }
-            }
-            println!();
-            println!("To install to VSCode: b00t-cli vscode install mcp <name>");
-            println!("To install to Claude Code: b00t-cli claude-code install mcp <name>");
-        }
-    }
-
-    Ok(())
-}
-
-fn get_mcp_config(name: &str, path: &str) -> Result<BootDatum> {
-    let mut path_buf = get_expanded_path(path)?;
-    path_buf.push(format!("{}.mcp.toml", name));
-
-    if !path_buf.exists() {
-        anyhow::bail!(
-            "MCP server '{}' not found. Use 'b00t-cli mcp add' to create it first.",
-            name
-        );
-    }
-
-    let content = fs::read_to_string(&path_buf).context(format!(
-        "Failed to read MCP config from {}",
-        path_buf.display()
-    ))?;
-
-    let config: UnifiedConfig =
-        toml::from_str(&content).context("Failed to parse MCP config TOML")?;
-
-    Ok(config.b00t)
-}
-
-fn vscode_install_mcp(name: &str, path: &str) -> Result<()> {
-    let datum = get_mcp_config(name, path)?;
-
-    let vscode_json = serde_json::json!({
-        "name": datum.name,
-        "command": datum.command.as_ref().unwrap_or(&"npx".to_string()),
-        "args": datum.args.as_ref().unwrap_or(&vec![])
-    });
-
-    let json_str =
-        serde_json::to_string(&vscode_json).context("Failed to serialize JSON for VSCode")?;
-
-    let result = cmd!("code", "--add-mcp", &json_str).run();
-
-    match result {
-        Ok(_) => {
-            println!(
-                "Successfully installed MCP server '{}' to VSCode",
-                datum.name
-            );
-            println!("VSCode command: code --add-mcp '{}'", json_str);
-        }
-        Err(e) => {
-            eprintln!("Failed to install MCP server to VSCode: {}", e);
-            eprintln!("Manual command: code --add-mcp '{}'", json_str);
-            std::process::exit(1);
-        }
-    }
-
-    Ok(())
-}
-
-fn claude_code_install_mcp(name: &str, path: &str) -> Result<()> {
-    let datum = get_mcp_config(name, path)?;
-
-    // Claude Code uses claude-code config add-mcp command
-    let claude_json = serde_json::json!({
-        "name": datum.name,
-        "command": datum.command.as_ref().unwrap_or(&"npx".to_string()),
-        "args": datum.args.as_ref().unwrap_or(&vec![])
-    });
-
-    let json_str =
-        serde_json::to_string(&claude_json).context("Failed to serialize JSON for Claude Code")?;
-
-    let result = cmd!("claude", "mcp", "add-json", &datum.name, &json_str).run();
-
-    match result {
-        Ok(_) => {
-            println!(
-                "Successfully installed MCP server '{}' to Claude Code",
-                datum.name
-            );
-            println!(
-                "Claude Code command: claude mcp add-json {} '{}'",
-                datum.name, json_str
-            );
-        }
-        Err(e) => {
-            eprintln!("Failed to install MCP server to Claude Code: {}", e);
-            eprintln!(
-                "Manual command: claude mcp add-json {} '{}'",
-                datum.name, json_str
-            );
-            std::process::exit(1);
-        }
-    }
-
-    Ok(())
-}
-
-fn cli_run(name: &str, path: &str) -> Result<()> {
-    let datum = get_cli_config(name, path)?;
-
-    if let Some(command) = &datum.command {
-        println!("Running CLI script '{}'...", name);
-        let result = cmd!("bash", "-c", command).run();
-        match result {
-            Ok(_) => {
-                println!("CLI script '{}' completed successfully.", name);
-            }
-            Err(e) => {
-                eprintln!("CLI script '{}' failed: {}", name, e);
-                std::process::exit(1);
-            }
-        }
-    } else {
-        eprintln!("No command defined for CLI script '{}'.", name);
-        std::process::exit(1);
-    }
-
-    Ok(())
-}
-
-fn cli_check(command: &str, path: &str) {
-    let desired_version_str = match get_cli_unified_config(command, path) {
-        Ok((config, _)) => config.b00t.desires.clone().unwrap_or_else(|| {
-            eprintln!("No desired version specified for {}", command);
-            std::process::exit(1);
-        }),
-        Err(e) => {
-            eprintln!("Error reading config for {}: {}", command, e);
-            std::process::exit(1);
-        }
-    };
-
-    let desired_version = Version::parse(&desired_version_str).unwrap();
-
-    if let Some(installed_version_str) = get_cli_installed_version(command, path) {
-        let installed_version = Version::parse(&installed_version_str).unwrap();
-
-        if installed_version == desired_version {
-            println!("🥾👍🏻 {}", command);
-            std::process::exit(0);
-        } else if installed_version > desired_version {
-            println!("🥾🐣 {} {}", command, installed_version);
-            std::process::exit(0);
-        } else {
-            println!(
-                "🥾😭 {} IS {} WANTS {}",
-                command, installed_version, desired_version
-            );
-            std::process::exit(1);
-        }
-    } else {
-        println!("🥾😱 {} MISSING", command);
-        std::process::exit(2);
-    }
-}
-
-fn get_cli_installed_version(command: &str, path: &str) -> Option<String> {
-    if let Ok((config, _)) = get_cli_unified_config(command, path) {
-        if let Some(version_cmd) = &config.b00t.version {
-            if let Ok(output) = cmd!("bash", "-c", version_cmd).read() {
-                let re = Regex::new(
-                    config
-                        .b00t
-                        .version_regex
-                        .as_deref()
-                        .unwrap_or("\\d+\\.\\d+\\.\\d+"),
-                )
-                .unwrap();
-                if let Some(caps) = re.captures(&output) {
-                    return Some(caps[0].to_string());
-                }
-            }
-        }
-    }
-    None
-}
-
-fn get_cli_config(name: &str, path: &str) -> Result<BootDatum> {
-    let mut path_buf = PathBuf::new();
-    path_buf.push(shellexpand::tilde(path).to_string());
-    path_buf.push(format!("{}.cli.toml", name));
-
-    if !path_buf.exists() {
-        anyhow::bail!(
-            "CLI script '{}' not found. Create a {}.cli.toml file first.",
-            name,
-            name
-        );
-    }
-
-    let content = fs::read_to_string(&path_buf).context(format!(
-        "Failed to read CLI config from {}",
-        path_buf.display()
-    ))?;
-
-    let config: UnifiedConfig =
-        toml::from_str(&content).context("Failed to parse CLI config TOML")?;
-
-    Ok(config.b00t)
-}
-
-fn cli_detect(command: &str, path: &str) {
-    if let Some(version) = get_cli_installed_version(command, path) {
-        println!("{}", version);
-    } else {
-        eprintln!("🥾😱 {} MISSING", command);
-        std::process::exit(2);
-    }
-}
-
-fn cli_desires(command: &str, path: &str) {
-    match get_cli_unified_config(command, path) {
-        Ok((config, _)) => {
-            println!("{}", config.b00t.desires.as_deref().unwrap_or("N/A"));
-        }
-        Err(e) => eprintln!("Error reading config for {}: {}", command, e),
-    }
-}
-
-fn cli_install(command: &str, path: &str) {
-    match get_cli_unified_config(command, path) {
-        Ok((config, filename)) => match config.b00t.get_datum_type(Some(&filename)) {
-            DatumType::Unknown => {
-                if let Some(install_cmd) = &config.b00t.install {
-                    println!("Installing {}...", command);
-                    let result = cmd!("bash", "-c", install_cmd).run();
-                    match result {
-                        Ok(_) => println!("Installation complete."),
-                        Err(e) => eprintln!("Installation failed: {}", e),
-                    }
-                } else {
-                    eprintln!("No install command defined for {}", command);
-                }
-            }
-            DatumType::Mcp => {
-                eprintln!(
-                    "Use 'b00t-cli vscode install mcp {}' or 'b00t-cli claude-code install mcp {}' instead",
-                    command, command
-                );
-            }
-            _ => {
-                eprintln!(
-                    "Installation not yet supported for {:?} packages",
-                    config.b00t.get_datum_type(Some(&filename))
-                );
-            }
-        },
-        Err(e) => eprintln!("Error reading config for {}: {}", command, e),
-    }
-}
-
-fn cli_update(command: &str, path: &str) {
-    match get_cli_unified_config(command, path) {
-        Ok((config, filename)) => match config.b00t.get_datum_type(Some(&filename)) {
-            DatumType::Unknown => {
-                if let Some(update_cmd) =
-                    config.b00t.update.as_ref().or(config.b00t.install.as_ref())
-                {
-                    println!("Updating {}...", command);
-                    let result = cmd!("bash", "-c", update_cmd).run();
-                    match result {
-                        Ok(_) => println!("Update complete."),
-                        Err(e) => eprintln!("Update failed: {}", e),
-                    }
-                } else {
-                    eprintln!("No update or install command defined for {}", command);
-                }
-            }
-            _ => {
-                eprintln!(
-                    "Update not yet supported for {:?} packages",
-                    config.b00t.get_datum_type(Some(&filename))
-                );
-            }
-        },
-        Err(e) => eprintln!("Error reading config for {}: {}", command, e),
-    }
-}
-
-fn cli_up(path: &str) {
-    let path_buf = PathBuf::from(shellexpand::tilde(path).to_string());
-    let entries = match fs::read_dir(&path_buf) {
-        Ok(entries) => entries,
-        Err(e) => {
-            eprintln!("Error reading directory {}: {}", path_buf.display(), e);
-            return;
-        }
-    };
-
-    for entry in entries {
-        if let Ok(entry) = entry {
-            let entry_path = entry.path();
-            if entry_path.extension().and_then(|s| s.to_str()) == Some("toml") {
-                let filename = entry_path
-                    .file_name()
-                    .and_then(|s| s.to_str())
-                    .unwrap_or("");
-                if filename.ends_with(".cli.toml") {
-                    if let Some(command) = entry_path.file_stem().and_then(|s| s.to_str()) {
-                        let command = command.trim_end_matches(".cli");
-                        let desired_version_str = match get_cli_unified_config(command, path) {
-                            Ok((config, _)) => match config.b00t.desires.clone() {
-                                Some(version) => version,
-                                None => continue,
-                            },
-                            Err(_) => continue,
-                        };
-
-                        if let Some(installed_version_str) =
-                            get_cli_installed_version(command, path)
-                        {
-                            if let (Ok(desired_version), Ok(installed_version)) = (
-                                semver::Version::parse(&desired_version_str),
-                                semver::Version::parse(&installed_version_str),
-                            ) {
-                                if installed_version < desired_version {
-                                    println!(
-                                        "Updating {} from {} to {}...",
-                                        command, installed_version, desired_version
-                                    );
-                                    cli_update(command, path);
-                                }
-                            }
-                        } else {
-                            println!("Installing missing package {}...", command);
-                            cli_install(command, path);
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn mcp_output(path: &str, use_mcp_servers_wrapper: bool, servers: &str) -> Result<()> {
-    let requested_servers: Vec<&str> = servers.split(',').map(|s| s.trim()).collect();
-    let mut server_configs = serde_json::Map::new();
-
-    for server_name in requested_servers {
-        if server_name.is_empty() {
-            continue;
-        }
-
-        match get_mcp_config(server_name, path) {
-            Ok(datum) => {
-                let mut server_config = serde_json::Map::new();
-                server_config.insert(
-                    "command".to_string(),
-                    serde_json::Value::String(datum.command.unwrap_or_else(|| "npx".to_string())),
-                );
-                server_config.insert(
-                    "args".to_string(),
-                    serde_json::Value::Array(
-                        datum
-                            .args
-                            .unwrap_or_default()
-                            .into_iter()
-                            .map(serde_json::Value::String)
-                            .collect(),
-                    ),
-                );
-
-                server_configs.insert(
-                    server_name.to_string(),
-                    serde_json::Value::Object(server_config),
-                );
-            }
-            Err(_) => {
-                // Create a cute poopy log error indicator instead of stderr warning
-                let mut error_config = serde_json::Map::new();
-                error_config.insert(
-                    "command".to_string(),
-                    serde_json::Value::String("b00t:💩🪵".to_string()),
-                );
-
-                let utc_timestamp = std::time::SystemTime::now()
-                    .duration_since(std::time::UNIX_EPOCH)
-                    .unwrap()
-                    .as_secs();
-                let utc_time = chrono::DateTime::from_timestamp(utc_timestamp as i64, 0)
-                    .unwrap()
-                    .format("%Y-%m-%dT%H:%M:%SZ")
-                    .to_string();
-
-                error_config.insert(
-                    "args".to_string(),
-                    serde_json::Value::Array(vec![
-                        serde_json::Value::String(utc_time),
-                        serde_json::Value::String(format!(
-                            "server '{}' not found in _b00t_ directory",
-                            server_name
-                        )),
-                    ]),
-                );
-
-                server_configs.insert(
-                    server_name.to_string(),
-                    serde_json::Value::Object(error_config),
-                );
-            }
-        }
-    }
-
-    let output = if use_mcp_servers_wrapper {
-        let mut wrapper = serde_json::Map::new();
-        wrapper.insert(
-            "mcpServers".to_string(),
-            serde_json::Value::Object(server_configs),
-        );
-        serde_json::Value::Object(wrapper)
-    } else {
-        serde_json::Value::Object(server_configs)
-    };
-
-    let json_str =
-        serde_json::to_string_pretty(&output).context("Failed to serialize MCP servers to JSON")?;
-    println!("{}", json_str);
-
-    Ok(())
-}
-
-fn get_ai_toml_files(path: &str) -> Result<Vec<String>> {
-    let expanded_path = get_expanded_path(path)?;
-    let entries = fs::read_dir(&expanded_path)
-        .with_context(|| format!("Error reading directory {}", expanded_path.display()))?;
-
-    let mut ai_files = Vec::new();
-    for entry in entries {
-        if let Ok(entry) = entry {
-            let entry_path = entry.path();
-            if let Some(file_name) = entry_path.file_name().and_then(|s| s.to_str()) {
-                if file_name.ends_with(".ai.toml") {
-                    if let Some(provider_name) = file_name.strip_suffix(".ai.toml") {
-                        ai_files.push(provider_name.to_string());
-                    }
-                }
-            }
-        }
-    }
-    Ok(ai_files)
-}
-
-fn get_ai_config(name: &str, path: &str) -> Result<AiConfig> {
-    let mut path_buf = get_expanded_path(path)?;
-    path_buf.push(format!("{}.ai.toml", name));
-
-    if !path_buf.exists() {
-        anyhow::bail!(
-            "AI provider '{}' not found. Use 'b00t-cli ai add' to create it first.",
-            name
-        );
-    }
-
-    let content = fs::read_to_string(&path_buf).context(format!(
-        "Failed to read AI config from {}",
-        path_buf.display()
-    ))?;
-
-    let config: AiConfig = toml::from_str(&content).context("Failed to parse AI config TOML")?;
-
-    Ok(config)
-}
-
-fn ai_add(file_path: &str, path: &str) -> Result<()> {
-    let content = fs::read_to_string(file_path)
-        .context(format!("Failed to read AI config file: {}", file_path))?;
-
-    let config: AiConfig = toml::from_str(&content).context("Failed to parse AI config TOML")?;
-
-    create_ai_toml_config(&config, path)?;
-
-    println!("AI provider '{}' configuration saved.", config.b00t.name);
-
-    Ok(())
-}
-
-fn ai_list(path: &str, json_output: bool) -> Result<()> {
-    let ai_files = get_ai_toml_files(path)?;
-    let mut ai_items = Vec::new();
-
-    for provider_name in ai_files {
-        match get_ai_config(&provider_name, path) {
-            Ok(config) => {
-                let model_names = config
-                    .models
-                    .as_ref()
-                    .map(|models| models.keys().cloned().collect())
-                    .unwrap_or_default();
-                let env_keys = config
-                    .env
-                    .as_ref()
-                    .map(|env| env.keys().cloned().collect())
-                    .unwrap_or_default();
-
-                ai_items.push(AiListItem {
-                    name: provider_name,
-                    models: Some(model_names),
-                    env_keys: Some(env_keys),
-                    error: None,
-                });
-            }
-            Err(e) => {
-                ai_items.push(AiListItem {
-                    name: provider_name,
-                    models: None,
-                    env_keys: None,
-                    error: Some(e.to_string()),
-                });
-            }
-        }
-    }
-
-    if json_output {
-        let expanded_path = get_expanded_path(path)?;
-        let output = AiListOutput {
-            providers: ai_items,
-            path: expanded_path.display().to_string(),
-        };
-        let json_str =
-            serde_json::to_string_pretty(&output).context("Failed to serialize AI list to JSON")?;
-        println!("{}", json_str);
-    } else {
-        let expanded_path = get_expanded_path(path)?;
-        if ai_items.is_empty() {
-            println!(
-                "No AI provider configurations found in {}",
-                expanded_path.display()
-            );
-            println!("Use 'b00t-cli ai add <file>' to add AI provider configurations.");
-        } else {
-            println!("Available AI providers in {}:", expanded_path.display());
-            println!();
-            for item in ai_items {
-                match (&item.models, &item.env_keys) {
-                    (Some(models), Some(env_keys)) => {
-                        println!(
-                            "🤖 {} ({} models, {} env vars)",
-                            item.name,
-                            models.len(),
-                            env_keys.len()
-                        );
-                        if !models.is_empty() {
-                            println!("   models: {}", models.join(", "));
-                        }
-                        if !env_keys.is_empty() {
-                            println!("   env: {}", env_keys.join(", "));
-                        }
-                    }
-                    _ => {
-                        println!("❌ {} (error reading config)", item.name);
-                    }
-                }
-            }
-            println!();
-            println!("Use 'b00t-cli ai output --kv <providers>' for environment variables");
-            println!("Use 'b00t-cli ai output --b00t <providers>' for TOML format");
-        }
-    }
-
-    Ok(())
-}
-
-fn substitute_env_vars(template: &str) -> String {
-    let mut result = template.to_string();
-    let env_var_pattern = Regex::new(r"\$\{([^}]+)\}").unwrap();
-
-    while let Some(captures) = env_var_pattern.captures(&result) {
-        let full_match = captures.get(0).unwrap().as_str();
-        let var_name = captures.get(1).unwrap().as_str();
-        let var_value = std::env::var(var_name).unwrap_or_default();
-        result = result.replace(full_match, &var_value);
-    }
-
-    result
-}
-
-fn ai_output(path: &str, format: &str, providers: &str) -> Result<()> {
-    let requested_providers: Vec<&str> = providers.split(',').map(|s| s.trim()).collect();
-
-    for provider_name in requested_providers {
-        if provider_name.is_empty() {
-            continue;
-        }
-
-        match get_ai_config(provider_name, path) {
-            Ok(config) => {
-                match format {
-                    "kv" => {
-                        if let Some(env_vars) = &config.env {
-                            for (key, value) in env_vars {
-                                let final_value = if value.is_empty() {
-                                    // If value is empty, try to get from environment
-                                    std::env::var(key).unwrap_or_default()
-                                } else {
-                                    // Manual environment variable substitution
-                                    substitute_env_vars(value)
-                                };
-                                println!("{}={}", key, final_value);
-                            }
-                        }
-                    }
-                    "b00t" | _ => {
-                        let toml_str = toml::to_string(&config)
-                            .context("Failed to serialize AI config to TOML")?;
-                        println!("{}", toml_str);
-                    }
-                }
-            }
-            Err(e) => {
-                eprintln!(
-                    "Warning: Failed to read config for '{}': {}",
-                    provider_name, e
-                );
-                continue;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn vscode_extension_list(json_output: bool) -> Result<()> {
-    let result = cmd!("code", "--list-extensions").read();
-
-    match result {
-        Ok(output) => {
-            if json_output {
-                let extensions: Vec<String> = output
-                    .lines()
-                    .map(|line| line.trim().to_string())
-                    .filter(|line| !line.is_empty())
-                    .collect();
-
-                let json_data = serde_json::json!({
-                    "extensions": extensions
-                });
-                let json_str = serde_json::to_string_pretty(&json_data)
-                    .context("Failed to serialize extensions to JSON")?;
-                println!("{}", json_str);
-            } else {
-                if output.trim().is_empty() {
-                    println!("No VS Code extensions installed");
-                } else {
-                    println!("Installed VS Code extensions:");
-                    for line in output.lines() {
-                        let extension = line.trim();
-                        if !extension.is_empty() {
-                            println!("📦 {}", extension);
-                        }
-                    }
-                }
-            }
-        }
-        Err(e) => {
-            anyhow::bail!(
-                "Failed to list VS Code extensions: {}. Make sure VS Code CLI is installed and available.",
-                e
-            );
-        }
-    }
-
-    Ok(())
-}
-
-fn vscode_extension_install(extension_id: &str) -> Result<()> {
-    println!("Installing VS Code extension: {}", extension_id);
-
-    let result = cmd!("code", "--install-extension", extension_id).run();
-
-    match result {
-        Ok(_) => {
-            println!("✅ Successfully installed extension: {}", extension_id);
-        }
-        Err(e) => {
-            anyhow::bail!(
-                "Failed to install VS Code extension '{}': {}. Check that the extension ID is correct.",
-                extension_id,
-                e
-            );
-        }
-    }
-
-    Ok(())
-}
-
-fn vscode_extension_uninstall(extension_id: &str) -> Result<()> {
-    println!("Uninstalling VS Code extension: {}", extension_id);
-
-    let result = cmd!("code", "--uninstall-extension", extension_id).run();
-
-    match result {
-        Ok(_) => {
-            println!("✅ Successfully uninstalled extension: {}", extension_id);
-        }
-        Err(e) => {
-            anyhow::bail!(
-                "Failed to uninstall VS Code extension '{}': {}. Check that the extension is installed.",
-                extension_id,
-                e
-            );
-        }
-    }
-
-    Ok(())
-}
-
-// K8s command handlers
-fn k8s_deploy(
-    from_dockerfile: Option<&str>,
-    from_compose: Option<&str>,
-    image: Option<&str>,
-    name: Option<&str>,
-    namespace: Option<&str>,
-    env: &[String],
-    _path: &str,
-) -> Result<()> {
-    println!("🚧 K8s deployment not yet implemented");
-
-    if let Some(dockerfile) = from_dockerfile {
-        println!("📄 Would deploy from Dockerfile: {}", dockerfile);
-    } else if let Some(compose) = from_compose {
-        println!("🐙 Would deploy from docker-compose: {}", compose);
-    } else if let Some(img) = image {
-        println!("📦 Would deploy from image: {}", img);
-    }
-
-    if let Some(pod_name) = name {
-        println!("🏷️  Pod name: {}", pod_name);
-    }
-
-    let ns = namespace.unwrap_or("default");
-    println!("🏠 Namespace: {}", ns);
-
-    if !env.is_empty() {
-        println!("🔧 Environment variables:");
-        for var in env {
-            println!("   {}", var);
-        }
-    }
-
-    Ok(())
-}
-
-fn k8s_deploy_mcp(
-    server: &str,
-    namespace: Option<&str>,
-    name: Option<&str>,
-    path: &str,
-) -> Result<()> {
-    use b00t_cli::k8s::{
-        client::K8sClient, resources::ContainerBuilder, resources::PodBuilder,
-        resources::ResourceUtils,
-    };
-    use tokio::runtime::Runtime;
-
-    let ns = namespace.unwrap_or("default");
-    let default_pod_name = format!("{}-mcp", server);
-    let pod_name = name.unwrap_or(&default_pod_name);
-
-    println!("🤖 Deploying MCP server to Kubernetes:");
-    println!("   Server: {}", server);
-    println!("   Pod name: {}", pod_name);
-    println!("   Namespace: {}", ns);
-
-    // Load MCP server configuration
-    let mcp_datum = match get_mcp_config(server, path) {
-        Ok(datum) => datum,
-        Err(e) => {
-            anyhow::bail!(
-                "Failed to load MCP server configuration for '{}': {}",
-                server,
-                e
-            );
-        }
-    };
-
-    println!("✅ Loaded MCP server configuration");
-
-    // Create async runtime
-    let rt = Runtime::new()?;
-
-    rt.block_on(async {
-        // Initialize k8s client
-        let client = match K8sClient::new().await {
-            Ok(client) => client,
-            Err(e) => {
-                anyhow::bail!("Failed to connect to Kubernetes cluster: {}", e);
-            }
-        };
-
-        println!("✅ Connected to Kubernetes cluster");
-
-        // Build container spec from MCP configuration
-        let mut container_builder = ContainerBuilder::new(
-            pod_name,
-            "node:18-alpine", // Default base image for MCP servers
-        );
-
-        // Set command and args from MCP config
-        if let Some(ref command) = mcp_datum.command {
-            container_builder = container_builder.command(vec![command.clone()]);
-        }
-
-        if let Some(ref args) = mcp_datum.args {
-            container_builder = container_builder.args(args.clone());
-        }
-
-        // Add environment variables
-        if let Some(ref env) = mcp_datum.env {
-            for (key, value) in env {
-                container_builder = container_builder.env(key, value);
-            }
-        }
-
-        // Add standard MCP port
-        container_builder = container_builder.port(3000, Some("mcp-port".to_string()));
-
-        let container = container_builder.build();
-
-        // Build pod spec
-        let labels = ResourceUtils::standard_labels(pod_name);
-        let pod = PodBuilder::new(pod_name)
-            .namespace(ns)
-            .labels(labels)
-            .container(container)
-            .restart_policy("Always")
-            .build()?;
-
-        // Deploy the pod
-        match client.deploy_pod(pod).await {
-            Ok(deployed_pod) => {
-                println!("🚀 Successfully deployed MCP server pod");
-                if let Some(pod_name) = deployed_pod.metadata.name {
-                    println!("   Pod name: {}", pod_name);
-                }
-                if let Some(pod_ns) = deployed_pod.metadata.namespace {
-                    println!("   Namespace: {}", pod_ns);
-                }
-                println!("💡 Use 'b00t-cli k8s logs {}' to view logs", pod_name);
-            }
-            Err(e) => {
-                anyhow::bail!("Failed to deploy MCP server pod: {}", e);
-            }
-        }
-
-        Ok::<(), anyhow::Error>(())
-    })?;
-
-    Ok(())
-}
-
-fn k8s_list(namespace: Option<&str>, all: bool, json: bool, _path: &str) -> Result<()> {
-    use b00t_cli::k8s::{client::K8sClient, utils::K8sUtils};
-    use tokio::runtime::Runtime;
-
-    let filter = if all { "all pods" } else { "b00t-managed pods" };
-    let ns = namespace.unwrap_or("all namespaces");
-
-    println!("📋 Listing {} in {}", filter, ns);
-
-    // Create async runtime
-    let rt = Runtime::new()?;
-
-    rt.block_on(async {
-        // Initialize k8s client
-        let client = match K8sClient::new().await {
-            Ok(client) => client,
-            Err(e) => {
-                anyhow::bail!("Failed to connect to Kubernetes cluster: {}", e);
-            }
-        };
-
-        // List pods
-        let pods = if all {
-            // List all pods in namespace or all namespaces
-            match client.list_all_pods(namespace).await {
-                Ok(pods) => pods,
-                Err(e) => {
-                    anyhow::bail!("Failed to list pods: {}", e);
-                }
-            }
-        } else {
-            // List only b00t-managed pods
-            match client.list_b00t_pods().await {
-                Ok(pods) => pods,
-                Err(e) => {
-                    anyhow::bail!("Failed to list b00t-managed pods: {}", e);
-                }
-            }
-        };
-
-        if json {
-            // Output as JSON
-            let pod_info: Vec<serde_json::Value> = pods
-                .iter()
-                .filter(|pod| all || K8sUtils::is_b00t_managed(pod))
-                .map(|pod| {
-                    serde_json::json!({
-                        "name": pod.metadata.name,
-                        "namespace": pod.metadata.namespace,
-                        "status": K8sUtils::format_pod_status(pod),
-                        "ready": K8sUtils::is_pod_ready(pod),
-                        "running": K8sUtils::is_pod_running(pod),
-                        "restarts": K8sUtils::get_restart_count(pod),
-                        "app": K8sUtils::extract_app_name(pod)
-                    })
-                })
-                .collect();
-
-            println!("{}", serde_json::to_string_pretty(&pod_info)?);
-        } else {
-            // Table output
-            let filtered_pods: Vec<_> = pods
-                .iter()
-                .filter(|pod| all || K8sUtils::is_b00t_managed(pod))
-                .collect();
-
-            if filtered_pods.is_empty() {
-                println!("No pods found");
-            } else {
-                println!("PODS:");
-                for pod in filtered_pods {
-                    let name = pod.metadata.name.as_deref().unwrap_or("unnamed");
-                    let pod_ns = pod.metadata.namespace.as_deref().unwrap_or("default");
-                    let status = K8sUtils::format_pod_status(pod);
-                    let app_name =
-                        K8sUtils::extract_app_name(pod).unwrap_or_else(|| "unknown".to_string());
-                    let ready_indicator = if K8sUtils::is_pod_ready(pod) {
-                        "✅"
-                    } else {
-                        "❌"
-                    };
-
-                    println!(
-                        "{} {:20} | {:15} | {:10} | {}",
-                        ready_indicator, name, status, pod_ns, app_name
-                    );
-                }
-            }
-        }
-
-        Ok::<(), anyhow::Error>(())
-    })?;
-
-    Ok(())
-}
-
-fn k8s_logs(
-    pod_name: &str,
-    namespace: Option<&str>,
-    follow: bool,
-    previous: bool,
-    tail: Option<i64>,
-    _path: &str,
-) -> Result<()> {
-    let ns = namespace.unwrap_or("default");
-
-    println!(
-        "📄 Getting logs for pod '{}' in namespace '{}'",
-        pod_name, ns
-    );
-    if follow {
-        println!("👁️  Following logs...");
-    }
-    if previous {
-        println!("⏮️  Showing previous container logs");
-    }
-    if let Some(lines) = tail {
-        println!("📏 Tail: {} lines", lines);
-    }
-    println!("🚧 Log retrieval not yet implemented");
-
-    Ok(())
-}
-
-fn k8s_delete(pod_name: &str, namespace: Option<&str>, _path: &str) -> Result<()> {
-    let ns = namespace.unwrap_or("default");
-
-    println!("🗑️  Deleting pod '{}' from namespace '{}'", pod_name, ns);
-    println!("🚧 Pod deletion not yet implemented");
-
-    Ok(())
-}
-
-fn k8s_restart_mcp(server_name: &str, namespace: Option<&str>, _path: &str) -> Result<()> {
-    let ns = namespace.unwrap_or("default");
-    let pod_name = format!("{}-mcp", server_name);
-
-    println!(
-        "🔄 Restarting MCP server pod '{}' in namespace '{}'",
-        pod_name, ns
-    );
-    println!("🚧 MCP restart not yet implemented");
-
-    Ok(())
-}
-
-fn k8s_status(detailed: bool, _path: &str) -> Result<()> {
-    println!("📊 Kubernetes cluster status:");
-    if detailed {
-        println!("🔍 Detailed status requested");
-    }
-    println!("🚧 Status check not yet implemented");
-
-    Ok(())
-}
-
-fn k8s_setup(force: bool, _path: &str) -> Result<()> {
-    println!("🚀 Setting up Kubernetes cluster...");
-    if force {
-        println!("💪 Force mode: will recreate existing cluster");
-    }
-    println!("🚧 Cluster setup not yet implemented");
-
-    Ok(())
-}
-
-fn k8s_teardown(all: bool, namespace: Option<&str>, _path: &str) -> Result<()> {
-    if all {
-        println!("💥 Tearing down entire cluster");
-    } else if let Some(ns) = namespace {
-        println!("🧹 Cleaning up namespace: {}", ns);
-    } else {
-        println!("🧹 Cleaning up default namespace");
-    }
-    println!("🚧 Teardown not yet implemented");
-
-    Ok(())
 }
